remove plotter and miner module

See merge request lianyu/bco!63

+++ src/miner.cpp
@@ -127,7 +127,7 @@ std::unique_ptr<CBlockTemplate> BlockAssembler::CreateNewBlock(const CScript& sc
     // The following code will make miners motivated to earlier post block. Maybe make it less difficulty.
     // pblock->nTime = GetAdjustedTime();
     //
-    if (GetAdjustedTime() > pindexPrev->GetBlockTime() + deadline + MAX_FUTURE_BLOCK_TIME) {
+    if (GetAdjustedTime() > (int64_t) (pindexPrev->GetBlockTime() + deadline + MAX_FUTURE_BLOCK_TIME)) {
         // Time changed
         pblock->nTime = static_cast<uint32_t>(GetAdjustedTime());
     } else {
